+++ src/components/CharacterCards.vue
@@ -41,32 +41,9 @@
 </template>
 
 <script setup>
-import axios from "axios";
-import { computed, ref } from "vue";
-import { orderBy } from "lodash";
+import { useFetchAllCharacters } from "./fetchAllCharacters.js";
 
-const loadingState = ref(null);
-const characters = ref([]);
-const orderKey = ref("id");
-
-const charactersOrdered = computed(() => {
-  return orderBy(characters.value, orderKey.value);
-});
-
-function setOrderKey(key) {
-  orderKey.value = key;
-}
-
-function fetchAllCharacters() {
-  loadingState.value = "loading";
-  axios.get("https://rickandmortyapi.com/api/character").then((response) => {
-    setTimeout(() => {
-      loadingState.value = "success";
-      characters.value = response.data.results;
-    }, 1000);
-  });
-}
-fetchAllCharacters();
+const { charactersOrdered, setOrderKey } = useFetchAllCharacters();
 </script>
 
 <style scoped>
